Refactor SpotlessTask into a legacy and non-legacy portion.
